+++ NEWS.md
@@ -1,3 +1,5 @@
+# paletteer (development version)
+
 # paletteer 1.4.0
 
 * `palettes_c_names` and `palettes_d_names` have been enhanced with more information.