+++ NEWS.md
@@ -1,4 +1,4 @@
-# paletteer (development version)
+# paletteer 1.5.0
 
 * Added sky palette from khroma package.
 * Added 31 Steven Universe palettes from tvthemes package.